refactor(winemaking-process): Refactor update endpoints

+++ ElixirLinePlatform.API/WinemakingProcess/Domain/Model/Commands/AddStage/AddClarificationStageCommand.cs
@@ -7,7 +7,7 @@ public record AddClarificationStageCommand(
     double turbidityAfterNtu,
     double volumeLiters,
     double temperature,
-    int durationHours,
+    double durationHours,
     string startedAt,
     string? completedBy,
     string? observations);
