+++ ElixirLinePlatform.API/WinemakingProcess/Domain/Model/Commands/UpdateStage/UpdateClarificationStageCommand.cs
@@ -10,5 +10,5 @@ public record UpdateClarificationStageCommand(
     double FinalTurbidityNtu,
     double WineVolumeLitres,
     double Temperature,
-    int DurationHours,
+    double DurationHours,
     string Observations);
